+++ src/main/java/com/maxtree/automotive/dashboard/view/admin/AssigningBusinessesToUserWindow.java
@@ -69,7 +69,7 @@ public class AssigningBusinessesToUserWindow extends Window {
 		List<Business> selected = new ArrayList<>();
 		for (Business b : allBusinesses) {
 			for (Business assignedBusiness : assignedBusinesses) {
-				if (assignedBusiness.getBusinessUniqueId() == b.getBusinessUniqueId()) {
+				if (assignedBusiness.getBusinessUniqueId().intValue() == b.getBusinessUniqueId().intValue()) {
 					selected.add(b);
 				}
 			}
